+++ lib/openapi_first/definition.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'failure'
-require_relative 'router'
-require_relative 'request'
-require_relative 'response'
 require_relative 'builder'
 require 'forwardable'