+++ lib/openapi_first/schema/validation_result.rb
@@ -17,7 +17,6 @@ module OpenapiFirst
         @errors ||= @validation.map do |err|
           ValidationError.new(
             value: err['data'],
-            message: err['error'],
             data_pointer: err['data_pointer'],
             schema_pointer: err['schema_pointer'],
             type: err['type'],